Add plumbing for streaming and patching annoations
Arrow, Band, and LabelSet were missing signal connections to
`source.streaming` and `source.plumbing` necessary for these annotations
to update when their data sources are streamed or patched.

This PR adds the missing signal connections. When reasonable, a
`request_render` was moved to `set_data` (this is ok when there is not
any CSS/dom rendering path)