Allow datetime values for number specs by default (#6278)
* Allow datetime values for number specs by default

This PR adds a parameter accept_datetime to NumberSpec that defaults to
True. When enabled, NumberSpecs will accept and immediately convert
datetime values to serialized ms-since-epoch values. This allows glyphs
and annotations to be positioned using datetime values.

This also fixes a future warning from pandas refarding pd.tslib

Datetime/timedelta detection and conversion was moved to
bokeh.util.serializtion so that the code might be shared.

* try tzinfo=None for datetime objs

* make serialized values for built in datetimes match numpy datetime64 values

* migration note