Allow arbitrary nested lists in layout function (#6743)
* Create recursive _create_grid function for layout

To create an arbitrary list of grids we require a recursive function
call.

* Test layout function against documented return values

I have added tests for layout ensuring that it returns a column of
`Row` layouts. This has been extended for the arbitrary layout to return
alternating Row and Column layouts.

There is also tests to ensure that all the children in the layout have
the same sizing_mode, and that it is as specified. All values of
bokeh.core.enums.SizingMode are tested.

* Remove unused variable