Issue #6704: Fix hover hittest on horizontal/vertical segments
Horizontal/vertical segments have zero bbox area and hence
are almost impossible to coincide with a point for hittest search.
Hence RBush.indices() was never returning such segments.
We now instead search rbush with a square of size 4x4 instead of a
point to find such segments as well.
Ideally, the square size should be maximum of the the line_width,
but the current state is in itself a great improvement.