Permit ColumnsStreamedEvent to accept DataFrames
When calling CDS stream with DataFrames, the DataFrame would make it all
the way to the ColumnsStreamedEvent, which presented it as-is for JSON
encoding. This fails because DataFrames are not directly JSON
serializable.

For compartmentalization and ease of testing, the issue is resolved by
teaching ColumnsStreamedEvent how to convert DataFrames to the
appropriate serializable dict itself. Other possible solutions not
pursued at this time:

* ensure ColumnsStreamedEvent only ever receives a dict (higher level)
* teach the Bokeh JSON encoder how to handle DataFrames (lower level)