PRO
Photo Recognition Optimization! (PRO) Out with the old in with the new. Think about it? Instead of (SEO) we'll have (PRO). You'll never have to type in a search bar again. Our DIRECT PHOTO RECOGNITION design (DPR) Examples; You're hiking, you're in a new state park, you observe a plant, an animal, or even a rock, which you have no idea what it is and you're curious. What we're introducing is an algortham which will immediately identify the image and define exactly what it is. This is just a small example. You're at a restaurant their's a salt and pepper shaker on the table and you absolutely love it, (DPR) will automatically examine the picture and route you to the cheapest site for purchasing. These examples hold true for clothing lines, car parts, building materials, and much more. Imagine the possibilities!!