Resolve pytest --log-file conflict
Recent versions of pytest fail with this error:
```
argparse.ArgumentError: argument --log-file: conflicting option string: --log-file
```

This PR changes the cusotm option added in tests/conftest.py to be `--examples-log-file`. A grep of the CI tests did not find any usage of the old value (the default value "examples.log" was used).