Fix errant scroll to end on DataTable with no selection (#7361)
* Fix errant scroll to end on DataTable with no selection

This problem was caused by applying the scroll computation in the case where there was an empty selection. In this instance, the scroll index was computed using `Math.min.apply(null, [])` which is `Infinity`, resulting in a scroll to the very end. The moves the logic for computing the index in to a method on the model and places it under test.

* use better unpacking syntax