Correct datetime conversion when subsecond fractions are present (#7366)
The existing code called `total_seconds` on a TimeDelta, as well as explicitly added `microseconds` back in. However, this resulted in a doubling of the microsecond component, since `total_seconds` already includes this portion:

https://docs.python.org/2/library/datetime.html#datetime.timedelta.total_seconds

This was never discoverd because all tests of `convert_datetime_type` on datetime values only exercised "exact date" values.