Address leaking memory (#7604)
* Address leaking memory

This PR adds a `destroy` method to Document, that removes several places where there might be circular references. However, this alone does not (yet) solve the problem, and a call to `gc.collect()` is required to fully release memory. A fuller solution down the road might be to implement a `destroy` protocol on all models so that they can be responsible for cleaning up any cycles.

* add memory log option

* can't format docstings of standard py2 classes