Stops the directory handler from bailing inappropriately.
At the start of it's modify_doc, the directory handler checked it's `failed` property and returned immediately if it was true. The `failed` property checked the `failed` properties of the embedded `_main_handler` and `_lifecycle_handler`. Checking `_main_handler.failed` and bailing immediately meant that any failure in the user's script would never be cleared, and the bokeh server would have to be restarted. Now we only check `_lifecycle_handler.failed`, which we do want to bail on.