Use BokehJS from CDN when the commit is tagged (#9150)
Get the name of the version from the git tag instead of using Read the
Docs version's name.

This fixes the problem when the commit is a tagged and Read the Docs
is building `latest` version for example. With this commit, we are
setting the CDN version to use as the tag name instead of the Read the
Docs version's name, which is the correct name.