Improve error message if import of channels fails.
Currently channels in only availeble on pip, but not conda.
Thus the conda install mentioned in the developer guide will not install it.
To run pytest -m codebase (mentioned in the developer guide) to be added as
git pre-commit hook will fail, giving first-time-contributers a bad experience

Point the users to install it via pip instead