ClientSession: Add different error messages (Close #3653)
As mentioned in issue #3653 the error messages from
ClientSession.pull and ClientSession.push did not take the cause of the
connection loss into account.

Therefore we add different Disconnected States to the client/states.py
to distinguish the reason of the connection loss and store information
about the error itself (error id, error message). Those information is
accessible through the ClientConnection class and is used in the
ClientSession to create more descriptive error messages.

Co-authored-by: Dennis Mehrhoff <dennis.mehrhoff@web.de>