Descriptive error (#10254)
* ClientSession: Add different error messages (Close #3653)

As mentioned in issue #3653 the error messages from
ClientSession.pull and ClientSession.push did not take the cause of the
connection loss into account.

Therefore we add different Disconnected States to the client/states.py
to distinguish the reason of the connection loss and store information
about the error itself (error id, error message). Those information is
accessible through the ClientConnection class and is used in the
ClientSession to create more descriptive error messages.

Co-authored-by: Dennis Mehrhoff <dennis.mehrhoff@web.de>

* Fix Codebase checks.

* Removed Disconnected Sub-classes

Distinguishing the error reason is now realised by checking an enum
field in the DISCONNECTED-State. Furthermore is_http_error and
is_network_error are removed from the ClientConnection and
ClientSession classes, in favor of a property, which returns said enum
value (error_reason).

Co-authored-by: Dennis Mehrhoff <dennis.mehrhoff@web.de>

* Update docstrings and set error code of NO_ERROR to None

- Cut off the docstring from col 80
- Remove commented out code

Co-authored-by: Soenke Florian <soenke.florian@gmx.de>

* Remove commented code

Co-authored-by: Dennis Mehrhoff <dennis.mehrhoff@web.de>
Co-authored-by: Bryan Van de Ven <bryan@bokeh.org>