AutocompleteInput: allow user-defined value
One of the main purposes of the AutocompleteInput widget is to select a
value from a long list (for short lists, one could use Select). However,
lists tend to become long when people keep adding to them.

Previously, AutocompleteInput would not accept text input that was not
present in the list of completions, thus excluding users from the
process of building the list.

This commit adds a `restrict_value` option (default: True), which allows
developers to lift the restriction of values to the list of completions.