10539 graph.rst doc edit
Finishied editing graph.rst. Let me kno if you'd like me to make any
changes to it.

One thing to note is that it might be better to merge the code snippet
in the 'Edge and node renderers' section with the code example linked in
the 'Layout providers' section and re-word the sections to account for
this change. The reason for this is that it looks like the former section
provides a code sample that's lacking in functionality. The catch is of
course that there would be a lot more going on in one code snippet and the
section explaining it.