#10539 graph.rst doc edit (#10579)
* Thorough edit of the chapter on Jupyter notebooks.

* Minor indentation fix.

* 10539 graph.rst doc edit

Finishied editing graph.rst. Let me kno if you'd like me to make any
changes to it.

One thing to note is that it might be better to merge the code snippet
in the 'Edge and node renderers' section with the code example linked in
the 'Layout providers' section and re-word the sections to account for
this change. The reason for this is that it looks like the former section
provides a code sample that's lacking in functionality. The catch is of
course that there would be a lot more going on in one code snippet and the
section explaining it.

* Revert "Minor indentation fix."

This reverts commit 4a92e8e0f507a9d73ee4fcb27bef28314b723ee8.

* Revert "Thorough edit of the chapter on Jupyter notebooks."

This reverts commit b72fa27eeebfa2652ab70f3ff14295e0c6a7a5f9.

* Revert "Thorough edit of the chapter on Jupyter notebooks."

This reverts commit b72fa27eeebfa2652ab70f3ff14295e0c6a7a5f9.

* Merging of sections on edges and nodes and on layout providers

* Revision of the 'Edge and node renderers' section

* whitespace

* Fixed comments

Co-authored-by: Bryan Van de Ven <bryan@bokeh.org>