Fix full-circle wedge hit test (#12124)
When Wedge and AnnularWedge glyphs go all the way around a full circle,
the hit test would always fail. This is because `angle_between()`
normalizes between 0 and 2*pi, then returns false if they are equal.
This adds an additional check in the glyph's hit testing such that if
the start and end angle are more than a full circle apart, it is
considered a hit because the wedge is going all the way around in a
circle.