Add bokehjs directory to MANIFEST.in
In the Python ecosystem most of the times a `wheel` distribution is
created from the `sdist` distribution instead of the source tree
directly. This process is described by PEP 517.

Therefore, the original JS/TS files need to be present in the `sdist`
so they can compiled/build/bundled when creating the `wheel`, which
means that we have to include the `bokehjs` directory in the
`MANIFEST.in`.

There are other alternatives to this approach, for example the JS/TS
files could be compiles/build/bundled when creating the `sdist` and just
copied over to the `wheel`. This would require creating a modified
version of the `sdist` command.

It is very likely that adjusts in the `MANIFEST.in` file or
via `include/exclude_package_data` configuration, to get the proper
files included in the distribution archives.