Promote note to warning
Ref. #2367.

That the glyphs for `nonselection_glyph` and `hover_glyph` is not affected by size, etc. bit me in the back, and apparently has confused many others as well, judging from the issue tracker. I think this behaviour is very unexpected, and deserves a warning level note.

Also, I think it would be great to be more explicit about what *visual* properties are, or link to where to find that. For me, it is for example not obvious that size is not a visual property. I would love to to it myself, but I am not familiar enough with Bokeh to do that.