Add `examples/plotting/css_classes_themed.py`
This is a variation of `css_classes.py` example. In the original,
stylesheets are applied to individual models. Here, a "global"
stylesheet is added via a theme. The effect at runtime is the
same in both cases, but the later may be more appealing for
users migrating from using pre shadow DOM global stylesheets.