Address PR comments
This commit changes the name of `rx[x|1]` and `ry[0|1]` to `dx[0|1]` and
`dy[0|1]`. It also attempts to fix the cursor not aligning with the
drawing of a box in the `hbar` and `vbar` cases. It fixes it in one
direction, but not in the other, as I think we need the cursor
coordinates in order to create a check.