Fix scoping issue for box edit tool coordinates (#13501)
* Rename x & y coordinates

This commit modifies the name of the x & y coordinates that use the box
edit tool, as the original name collided with a few glyph types.

* Address PR comments

This commit changes the name of `rx[x|1]` and `ry[0|1]` to `dx[0|1]` and
`dy[0|1]`. It also attempts to fix the cursor not aligning with the
drawing of a box in the `hbar` and `vbar` cases. It fixes it in one
direction, but not in the other, as I think we need the cursor
coordinates in order to create a check.

* Fix HBar/VBar coordinates

* Allow setting key modifiers in PlotActions

* Deal with throttling threshold

* Add unit regression tests

---------

Co-authored-by: Mateusz Paprocki <mattpap@gmail.com>