Update util.py
Added reuse_port option to bokeh.server.util  bind_sockets, to allow for this behavior in a bokeh emded context.  Otherwise, when requesting a specific port from within a Flask app, we get an error "OSError: [Errno 98] Address already in use".  
The specific port is required to most efficiently manage open ports in a container environment.